+++ src/cookies/setCookie.ts
@@ -11,6 +11,7 @@ const setCookie = ({ name, value, expirationDays }: SetCookieProps) => {
     date.setTime(date.getTime() + expirationDays * 24 * 60 * 60 * 1000);
     expires = `; expires=${date.toUTCString()}`;
   }
+  // biome-ignore lint/suspicious/noDocumentCookie: Still works
   document.cookie = `${name}=${value}${expires}; path=/`;
 };